fix: Input 컴포넌트 통합 및 경로 수정
