+++ README.md
@@ -293,3 +293,8 @@
     - 기상청 날씨 데이터를 기반으로 날씨 12개로 상세 구분
     - 현재 날씨를 기반으로 아이콘 동적으로 출력
     - 세분화된 날씨 아이콘 업로드
+
+### 9월 10일
+
+- 황선우
+  - Input 컴포넌트 통합 및 경로 수정
